Started Basic Enemy AI
Created a Dummy Navmesh in PlayerTestScene

Added Unity AI package
Added Player Tag

Created EnemyAI script that allows you to set a navmeshagent destination.
Created TestEnemyAI to set destination to player every 0.5 seconds in a coroutine.

PlayerManager is now a Singleton that takes in PlayerController & can return it's position for things that want to find where the player is.

Added Spritefire model as a dummy NPC model