Blocked Out Rune Placement & Storage Logic
Blocked Out Rune Placement & Storage Logic
RuneAnchor is a component used to physically hold a rune gameobject on a slot.
RunePlaceManager handles the movement of runes around the inventory.
RunePlaceSubscriber is put on the slots (been made into buttons) to tell RunePlaceManager that this rune slot wants a rune placed in it.
RunePlaceInvoker starts the rune place sequence by clicking on an unassigned rune in your inventory.

Replaced Rune & Keystone Images